Sources/S5i-Projet-Client/pyproject.toml
[tool.poetry]
name = "s5i-projet-client"
version = "0.1.0"
description = ""
authors = ["Mathieu Sevegny <sevegny.mathieu@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
